Starting DOTS removal
DOTS has proved to be very complex to integrate with OOP. I've decided
to remove it for now. It can be added back whenever.

This creates a central VoraUtils namespace for all of these functions.
It includes an Astar class. This commit prepares for the complete
removal of DOTS and the substitution with steering behaviors.

+++ Unity_Project/Assets/Scripts/Generation/Board.cs
@@ -95,8 +95,8 @@ public class Board : MonoBehaviour {
 	public HashSet<Coordinate> GetShortestPath(Coordinate start,
 		Coordinate end,
 		Func<Coordinate, Coordinate, Coordinate[]> GetSuccessorsFunction,
-		Func<Coordinate, Coordinate, Coordinate, int, int> GetCostFunction,
-		Func<Coordinate, Coordinate, int> GetHeuristicFunction) {
+		Func<Coordinate, Coordinate, Coordinate, float, float> GetCostFunction,
+		Func<Coordinate, Coordinate, float> GetHeuristicFunction) {
 		return floor.GetShortestPath(start, end, GetSuccessorsFunction, GetCostFunction, GetHeuristicFunction);
 	}
 }
